TINS-496 not all overloads calculated
if a binding was already created, then we checked if a variable is fixed, if not we reported that at least one binding was created. This lead to the ambiguous overloads. Now we consider whether the variable is a real variable or an expression variable. In case of an expression variable we count it as a created variable binding. In case of a real variable we do not count it as a created variable binding.